requirements.txt
bumpversion==0.5.3
wheel==0.23.0
flake8==2.4.1
coverage==4.0
pytest==2.9.1
cryptography==1.0.1
PyYAML==3.11

